Move git parsing logic to shared file
